Add input reading framework
Previously I used inlined heredocs for input. This has some advantages:
* It's very easy to do while you're solving.
* It lets you look at your input without navigating to a different file.

But it has some disadvantages:
* If you want to run the task with a different input you have to edit
the code.
* Most tasks have a huge (real) input so it just clogs your screen and
makes line numbers useless.

A nice middle ground between the two is to have the sample input as part
of the code, while the real input is provided externally:
* The sample input is usually tiny and readable. Hence you get an idea
of the input's structure without getting overwhelmed.
* In order to run the task against a different input you just have to
replace a txt file.