Day 17 solutions
The second part requires multiple re-runs of the program with some changes in between. I didn't care to clean it up and make it working.

The basic idea is that you would have some rock structure at the start (head) than a repeating cycle of the same structure than some rock structure at the end (tail).

The head occurs because you start with a solid floor, whereas you're not guaranteed to get a solid row with your input.
The tail occurs because the last cycle will (likely) be incomplete.

Run once to find the cycle and the head. Calculate the size of the head + the size of the cycle * max amount of cycles.
Then run a second time starting just before the first cycle starts and run for howmany rocks remain. This way you can get the size of the tail.

Add the three together.